Splitted forgotten temples / fixed walking reduction
