+++ d2bs/kolbot/libs/bots/AutoPlay/modules/Dungeons/HighLevelForgottenTemples.js
@@ -2,49 +2,9 @@
 
 	const CustomDungeon = require('../CustomDungeon');
 	const walkTo = require('../WalkTo');
-	const Misc = require('../../../../modules/Misc');
 	const Pather = require('../../../../modules/Pather');
-	const Pickit = require('../../../../modules/Pickit');
 
 	// const FastestPath = require('../FastestPath');
-
-	new CustomDungeon(sdk.areas.RuinedTemple, function () {
-		const ps = getPresetUnit(sdk.areas.RuinedTemple, 2, sdk.units.LamEsensTome);
-
-		walkTo(ps, false);
-
-		if (!me.getQuest(sdk.quests.LamEsensTome, 0)) {
-			const LamEsen = ps.unit;
-			Misc.poll(() => {
-				LamEsen.click();
-				return LamEsen.mode;
-			});
-			Pickit.pickItems();
-		}
-
-	});
-
-	//ToDo; figure out if its a static map
-	new CustomDungeon(sdk.areas.DisusedFane, function () {
-		walkTo([
-			{x: 15024, y: 5640},
-			{x: 15025, y: 5692},
-			{x: 15059, y: 5695},
-			{x: 15097, y: 5704},
-			{x: 15082, y: 5673},
-			{x: 15057, y: 5634},
-		], false);
-	});
-
-	new CustomDungeon(sdk.areas.ForgottenReliquary, function () {
-		walkTo([
-			{x: 15089, y: 6294},
-			{x: 15085, y: 6345},
-			{x: 15030, y: 6302},
-			{x: 15024, y: 6340},
-		], false);
-	});
-
 
 	const chestIds = [5, 6, 87, 104, 105, 106, 107, 143, 140, 141, 144, 146, 147, 148, 176, 177, 181, 183, 198, 240, 241, 242, 243, 329, 330, 331, 332, 333, 334, 335, 336, 354, 355, 356, 371, 387, 389, 390, 391, 397, 405, 406, 407, 413, 420, 424, 425, 430, 431, 432, 433, 454, 455, 501, 502, 504, 505, 580, 581];
 
@@ -61,7 +21,6 @@
 
 		walkTo(getAllChestsAndSort(), false);
 	});
-
 
 	new CustomDungeon(sdk.areas.RuinedFane, function () {
